requirements.txt
transformers==4.41.0
datasets==2.19.1
accelerate==0.30.1
sentencepiece==0.2.0
bitsandbytes==0.42.0

scikit-learn==1.5.0
spacy==3.7.4
sentence-transformers==2.7.0
streamlit==1.35.0
pytest==8.2.0
requests==2.31.0
trafilatura==1.6.4
typer[all]==0.9.0
rich==13.7.0
pdfplumber==0.11.0
python-docx==1.1.0
tqdm==4.66.4
typer>=0.8.0
click>=8.0.0   
reportlab>=3.6.12
fpdf2 >= 1.0.0
cloudscraper >= 1.2.0
playwright
bs4